huy-update-2/11_lan_1

+++ medical_management/src/main/java/com/example/medical_management/repository/medical/ICategoryRepository.java
@@ -3,5 +3,5 @@ package com.example.medical_management.repository.medical;
 import com.example.medical_management.model.medical_supplies.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICategoryRepository extends JpaRepository<Category, Integer> {
+public interface ICategoryRepository extends JpaRepository<Category, Long> {
 }
